Don't flush on GST_SEGMENT_EVENT
In some cases, before a segment event arrives the decoder already
manages to queue some buffers up internally. If they're flushed
on a GST_EVENT_SEGMENT, there will be a noticeable video freeze
for a couple of seconds. It will last until the STC reaches PTS
received in the decoder after the bad flush.

This change also reverts a patchset added for a seperate issue,
which is why the scope is still being discussed:

https://jira.rdkcentral.com/jira/browse/BCMCZ-376