BCMCZ-379 Refactor position query handling
This commit refactors position query handling in the video decoder,
with the following changes:
    - recalculate_position function is removed, in favour of using
      a generic gstreamer function (`gst_segment_to_stream_time`),
    - query_position function is reworked, so that it now uses
      `gst_segment_to_stream_time` for absolute position calculation,
    - formatting fixes to break up long lines of code.

These changes also address an issue, where the video decoder did not
respond to position queries for a prolonged time during startup.
A position query should be handled, as soon as a segment event has
been processed.